# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter API",
    "version": "1.0.0",
    "description": "Clean API facade over the read-only Urja Meter Ops portal."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/meters": {
      "get": {
        "summary": "Search meters",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter search results"
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/energy": {
      "get": {
        "summary": "Get meter energy",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Energy readings"
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/location": {
      "get": {
        "summary": "Get meter location",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter location"
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "summary": "Get meter details",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meter details"
          },
          "404": {
            "description": "Meter not found"
          }
        }
      }
    }
  }
}